fixed bug in ClearableEditText where icon might not be displayed
At the investigation found that the readOnly attribute used in the
BindingAdapter is activated after setting the value in the
ClearableEditText widget.

So, the default value of the focusable attribute is true and the logic
of the ClearableEditText widget is that if the widget is focusable, the
icon should only be set if the widget has focus.